+++ packages/service-library/tests/redis/test_semaphore.py
@@ -7,7 +7,6 @@
 
 import asyncio
 import datetime
-from typing import Literal
 from unittest import mock
 
 import pytest
@@ -19,13 +18,11 @@ from servicelib.redis._constants import (
     SEMAPHORE_HOLDER_KEY_PREFIX,
     SEMAPHORE_KEY_PREFIX,
 )
-from servicelib.redis._errors import SemaphoreLostError
 from servicelib.redis._semaphore import (
     DistributedSemaphore,
     SemaphoreAcquisitionError,
     SemaphoreNotAcquiredError,
 )
-from servicelib.redis._semaphore_decorator import with_limited_concurrency
 
 pytest_simcore_core_services_selection = [
     "redis",
@@ -33,21 +30,6 @@ pytest_simcore_core_services_selection = [
 pytest_simcore_ops_services_selection = [
     "redis-commander",
 ]
-
-
-@pytest.fixture
-def semaphore_name(faker: Faker) -> str:
-    return faker.pystr()
-
-
-@pytest.fixture
-def semaphore_capacity() -> int:
-    return 3
-
-
-@pytest.fixture
-def short_ttl() -> datetime.timedelta:
-    return datetime.timedelta(seconds=1)
 
 
 @pytest.fixture
@@ -304,96 +286,6 @@ async def test_semaphore_context_manager_with_exception(
     assert await captured_semaphore.get_current_count() == 0
 
 
-async def test_semaphore_auto_renewal_via_decorator(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-    semaphore_capacity: int,
-    short_ttl: datetime.timedelta,
-):
-    """Test auto-renewal functionality through the decorator (TaskGroup approach)"""
-    work_started = asyncio.Event()
-    work_completed = asyncio.Event()
-
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=semaphore_capacity,
-        ttl=short_ttl,
-    )
-    async def long_running_work() -> Literal["success"]:
-        work_started.set()
-        # Wait longer than TTL to ensure renewal works
-        await asyncio.sleep(short_ttl.total_seconds() * 2)
-        work_completed.set()
-        return "success"
-
-    # Start the work
-    task = asyncio.create_task(long_running_work())
-
-    # Wait for work to start
-    await work_started.wait()
-
-    # Check that semaphore is being held
-    temp_semaphore = DistributedSemaphore(
-        redis_client=redis_client_sdk,
-        key=semaphore_name,
-        capacity=semaphore_capacity,
-        ttl=short_ttl,
-    )
-    assert await temp_semaphore.get_current_count() == 1
-
-    # Wait for work to complete
-    result = await task
-    assert result == "success"
-    assert work_completed.is_set()
-
-    # After completion, semaphore should be released
-    assert await temp_semaphore.get_current_count() == 0
-
-
-async def test_decorator_auto_renewal_failure_propagation(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-    semaphore_capacity: int,
-    short_ttl: datetime.timedelta,
-):
-    """Test that auto-renewal failures properly propagate as exceptions in the decorator"""
-
-    work_started = asyncio.Event()
-
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=semaphore_capacity,
-        ttl=short_ttl,
-    )
-    async def work_that_should_fail() -> Literal["should not reach here"]:
-        work_started.set()
-        # Wait long enough for renewal to be attempted multiple times
-        await asyncio.sleep(short_ttl.total_seconds() * 4)
-        return "should not reach here"
-
-    # The decorator should propagate the renewal failure
-    task = asyncio.create_task(work_that_should_fail())
-    await work_started.wait()  # Wait for work to start
-
-    # Wait for the first renewal interval to pass
-    renewal_interval = short_ttl / 3
-    await asyncio.sleep(
-        renewal_interval.total_seconds() + 1
-    )  # Wait for renewal to happen
-
-    # Find and delete all holder keys for this semaphore
-    holder_keys = await redis_client_sdk.redis.keys(
-        f"{SEMAPHORE_HOLDER_KEY_PREFIX}{semaphore_name}:*"
-    )
-    assert holder_keys, "Holder keys should exist before deletion"
-    await redis_client_sdk.redis.delete(*holder_keys)
-
-    with pytest.raises(SemaphoreLostError):
-        await task  # This should raise the renewal failure exception
-
-
 async def test_semaphore_ttl_cleanup(
     redis_client_sdk: RedisClientSDK,
     semaphore_name: str,
@@ -429,173 +321,6 @@ async def test_semaphore_ttl_cleanup(
     # Verify expired entry was removed
     remaining = await redis_client_sdk.redis.zcard(semaphore.semaphore_key)
     assert remaining == 0
-
-
-async def test_decorator_basic_functionality(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-):
-    call_count = 0
-
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=1,
-    )
-    async def limited_function():
-        nonlocal call_count
-        call_count += 1
-        await asyncio.sleep(0.1)
-        return call_count
-
-    # Multiple concurrent calls
-    tasks = [asyncio.create_task(limited_function()) for _ in range(3)]
-    results = await asyncio.gather(*tasks)
-
-    # All should complete successfully
-    assert len(results) == 3
-    assert all(isinstance(r, int) for r in results)
-
-
-async def test_decorator_with_callable_parameters(
-    redis_client_sdk: RedisClientSDK,
-):
-    executed_keys = []
-
-    def get_redis_client(*args, **kwargs):
-        return redis_client_sdk
-
-    def get_key(user_id: str, resource: str) -> str:
-        return f"{user_id}-{resource}"
-
-    def get_capacity(user_id: str, resource: str) -> int:
-        return 2
-
-    @with_limited_concurrency(
-        get_redis_client,
-        key=get_key,
-        capacity=get_capacity,
-    )
-    async def process_user_resource(user_id: str, resource: str):
-        executed_keys.append(f"{user_id}-{resource}")
-        await asyncio.sleep(0.05)
-
-    # Test with different parameters
-    await asyncio.gather(
-        process_user_resource("user1", "wallet1"),
-        process_user_resource("user1", "wallet2"),
-        process_user_resource("user2", "wallet1"),
-    )
-
-    assert len(executed_keys) == 3
-    assert "user1-wallet1" in executed_keys
-    assert "user1-wallet2" in executed_keys
-    assert "user2-wallet1" in executed_keys
-
-
-async def test_decorator_capacity_enforcement(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-):
-    concurrent_count = 0
-    max_concurrent = 0
-
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=2,
-    )
-    async def limited_function():
-        nonlocal concurrent_count, max_concurrent
-        concurrent_count += 1
-        max_concurrent = max(max_concurrent, concurrent_count)
-        await asyncio.sleep(0.1)
-        concurrent_count -= 1
-
-    # Start 5 concurrent tasks
-    tasks = [asyncio.create_task(limited_function()) for _ in range(5)]
-    await asyncio.gather(*tasks)
-
-    # Should never exceed capacity of 2
-    assert max_concurrent <= 2
-
-
-async def test_decorator_with_exception_handling(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-):
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=1,
-    )
-    async def failing_function():
-        raise RuntimeError("Test exception")
-
-    with pytest.raises(RuntimeError, match="Test exception"):
-        await failing_function()
-
-    # Semaphore should be released even after exception
-    # Test by trying to acquire again
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=1,
-    )
-    async def success_function():
-        return "success"
-
-    result = await success_function()
-    assert result == "success"
-
-
-async def test_decorator_non_blocking_behavior(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-):
-    # Test the blocking timeout behavior
-    started_event = asyncio.Event()
-
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=1,
-        blocking=True,
-        blocking_timeout=datetime.timedelta(seconds=0.1),
-    )
-    async def limited_function():
-        started_event.set()
-        await asyncio.sleep(0.2)
-
-    # Start first task that will hold the semaphore
-    task1 = asyncio.create_task(limited_function())
-    await started_event.wait()  # Wait until semaphore is actually acquired
-
-    # Second task should timeout and raise an exception
-    with pytest.raises(SemaphoreAcquisitionError):
-        await limited_function()
-
-    await task1
-
-
-async def test_decorator_preserves_function_metadata(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-):
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=1,
-    )
-    async def documented_function(arg1: str, arg2: int) -> str:
-        """Test function with documentation"""
-        return f"{arg1}-{arg2}"
-
-    assert documented_function.__name__ == "documented_function"
-    assert documented_function.__doc__ == "Test function with documentation"
-
-    result = await documented_function("test", 42)
-    assert result == "test-42"
 
 
 async def test_redis_connection_failure_during_acquire(
@@ -636,149 +361,3 @@ async def test_multiple_semaphores_different_keys(
         ),
     ):
         ...
-
-
-async def test_decorator_exception_reraise_with_auto_renewal(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-    semaphore_capacity: int,
-    short_ttl: datetime.timedelta,
-):
-    """Test that user function exceptions are properly re-raised even with auto-renewal running"""
-
-    class UserFunctionError(Exception):
-        """Custom exception to ensure we're catching the right exception"""
-
-    work_started = asyncio.Event()
-    renewal_count = 0
-
-    # Track that auto-renewal is actually happening
-    original_renew = __import__(
-        "servicelib.redis._semaphore_decorator", fromlist=["_renew_semaphore_entry"]
-    )._renew_semaphore_entry
-
-    async def tracking_renew_semaphore_entry(semaphore):
-        nonlocal renewal_count
-        renewal_count += 1
-        await original_renew(semaphore)
-
-    with mock.patch(
-        "servicelib.redis._semaphore_decorator._renew_semaphore_entry",
-        side_effect=tracking_renew_semaphore_entry,
-    ):
-
-        @with_limited_concurrency(
-            redis_client_sdk,
-            key=semaphore_name,
-            capacity=semaphore_capacity,
-            ttl=short_ttl,  # Short TTL to ensure renewal happens
-        )
-        async def failing_function():
-            work_started.set()
-            # Wait long enough for at least one renewal to happen
-            await asyncio.sleep(short_ttl.total_seconds() * 0.8)
-            # Then raise our custom exception
-            raise UserFunctionError("User function failed intentionally")
-
-        # Verify the exception is properly re-raised
-        with pytest.raises(
-            UserFunctionError, match="User function failed intentionally"
-        ):
-            await failing_function()
-
-        # Ensure work actually started
-        assert work_started.is_set()
-
-        # Verify auto-renewal was working (at least one renewal should have happened)
-        assert renewal_count >= 1, "Auto-renewal should have been called at least once"
-
-    # Verify semaphore was properly released by trying to acquire it again
-    test_semaphore = DistributedSemaphore(
-        redis_client=redis_client_sdk,
-        key=semaphore_name,
-        capacity=semaphore_capacity,
-        ttl=short_ttl,
-    )
-    assert (
-        await test_semaphore.get_current_count() == 0
-    ), "Semaphore should be released after exception"
-
-
-async def test_decorator_exception_types_preserved(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-    semaphore_capacity: int,
-):
-    """Test that different exception types are preserved through the decorator"""
-
-    # Test ValueError
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=f"{semaphore_name}_ValueError",
-        capacity=semaphore_capacity,
-    )
-    async def function_raising_value_error():
-        raise ValueError("Invalid value")
-
-    with pytest.raises(ValueError) as exc_info:
-        await function_raising_value_error()
-    assert str(exc_info.value) == "Invalid value"
-
-    # Test TypeError
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=f"{semaphore_name}_TypeError",
-        capacity=semaphore_capacity,
-    )
-    async def function_raising_type_error():
-        raise TypeError("Wrong type")
-
-    with pytest.raises(TypeError) as exc_info:
-        await function_raising_type_error()
-    assert str(exc_info.value) == "Wrong type"
-
-    # Test KeyError
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=f"{semaphore_name}_KeyError",
-        capacity=semaphore_capacity,
-    )
-    async def function_raising_key_error():
-        raise KeyError("Missing key")
-
-    with pytest.raises(KeyError) as exc_info:
-        await function_raising_key_error()
-    assert str(exc_info.value) == "'Missing key'"  # KeyError adds quotes
-
-    # Test RuntimeError
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=f"{semaphore_name}_RuntimeError",
-        capacity=semaphore_capacity,
-    )
-    async def function_raising_runtime_error():
-        raise RuntimeError("Runtime error")
-
-    with pytest.raises(RuntimeError) as exc_info:
-        await function_raising_runtime_error()
-    assert str(exc_info.value) == "Runtime error"
-
-
-async def test_decorator_cancelled_error_preserved(
-    redis_client_sdk: RedisClientSDK,
-    semaphore_name: str,
-    semaphore_capacity: int,
-):
-    """Test that CancelledError is properly preserved through the decorator"""
-
-    @with_limited_concurrency(
-        redis_client_sdk,
-        key=semaphore_name,
-        capacity=semaphore_capacity,
-    )
-    async def function_raising_cancelled_error():
-        raise asyncio.CancelledError
-
-    # Verify CancelledError is preserved
-    with pytest.raises(asyncio.CancelledError):
-        await function_raising_cancelled_error()
